Extend an abstract Player class rather than implement interface
