+++ src/main/java/tictactoe/Main.java
@@ -5,8 +5,8 @@ public class Main {
     public static void main(String[] args) {
         Display display = new Display(System.out, System.in);
         Board board = new Board();
-        Player playerX = new Player("X", display);
-        Player playerO = new Player("O", display);
+        Player playerX = new HumanPlayer("X", display);
+        Player playerO = new HumanPlayer("O", display);
         Game game = new Game(display, board, playerX, playerO);
         Session session = new Session(game, display);
